ch02/pyproject.toml
[project]
name = "ch02"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.46.0",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "jupyterlab>=4.4.10",
    "langchain-community>=0.4.1",
    "pypdf>=6.1.3",
    "python-dotenv>=1.1.1",
]
